Enable Person to be created from couchdb or sql if the former does not exist
Reduce from create_entity to create_person, i.e., we only create person either from couchdb if possible, if not use EntityDatum table in sql.  We return the person, which can be either Person or EntityDatum, but both supports errors so it's easy to handle and both have id field that is used for Entity to link to.
We also changed validation of the Entity property_document_id to validate against either couchdb or EntityDatum serial number.
We generated scaffold for EntityData and migrated the db.