Added functional tests for inbound email from id-able source, moodTags to meantItRel, association between meantItRel with inbound_email.
Setup fixures for test/fixtures/inbound_email.yml.  This creates db entries but are separated from those we will create in the tests themselves.
Created moodTags for meantItRel, which is important for those sent to other@meant.it  We use some reasoner to classify them into mood and tag the meantItRel.  This is managed through meantItMoodTagRels table.