Allow nick to be null but only if pii is not otherwise throw error.
Created migration to drop nick column in end_points to be not null.
Added logic to handle the four cases where nick/pii can be empty/provided.
Make respond_to xml in 'def create' of inbound_emails_controller.rb return status 200 if the path is inbound_emails_200 so that sendgrid knows that it is successful instead of redirecting to a page to show the new entry which is useful only to a human user but not sendgrid.