Fixed issue on meant.it sent from web page not showing and two endpoints created for the same email pii.
We need to save both ways, i.e., after sender_pii.endpoint_id = sender_endpoint.id, we need to sender_pii.save and sender_endpoint.save to solve the first issue.  Otherwise, when we use sender_endpoint, it won't have pii.
We also made a mistake in checking that the endpoint of pii that has id = creator_endpoint_id is > 1, when it should be > 0, thus resulting in re-creation of a sender_endpoint even though there is already one.