Pii with caps failing to be recognized, auto entity association.
We used the original input_str to detect the location of pii, the former may contain mixed case while the latter has been downcased.  Create and used input_str_downcase.
Added code in inbound_emails_controller.rb to auto associate, i.e., find entity, if it exists, create sender_endpoint if it does not exist for the entity and link the entity to the sender_endpoint.