Enable entity to use same email for now
To do so we removed the unique index by_email.
Alternative is to enforce unique mail.  However, this will be troublesome for user since if their email is being used other others, they will have to claim it.  Without this unique enforcement, other people may use same email but they cannot read any messages or reply to anything sent to that email.  They best they can do is to trigger mails sent to it.