Create nick same as pii if pii is not email and pii not set hidden. Fix issue on confirmation email to claim pii pointing to localhost.
We often do ':hello_kitty cat with no mouth' and end up 'cat' becomes nick by virtue of being the first keyword that is not pii.  So when pii is not email and not hidden we use the pii as the nick as well, so in this case hello_kitty becomes the nick.
For email pii, using it as nick is no-no due to privacy and even if we just use the part prior the '@' since provide can be guessed or the prefix before '@' may not make much sense.  Besides user will be more aware of providing a nick.
The other problem is when we try to claim pii, the confirmation email is sent but contains http://localhost:3000... instead of http://meant.it.  Fix this by moving the config.action_mailer.default_url_options to both config/environments/development.rb and config/environments/production.rb.