Handled issue of incomplete confirmation link process.
A user may click on the completion link but may not be logged in or did not login resulting in confirmed_at field of the devise User being filled but not tied to any entity.
We check if pii is tied to entity when claiming; if it is not then we nullify confirm if the User already exists (indicator of incomplete confirmation process).
We have kinda of hard-coded that each pii is only tied to a single entity although in our model a pii can belongs_to many entities.  This is because of the single entry of devise User only.  We can keep on clearing the devise User to allow multiple confirmation but that makes things complex.
There is still issue of wrong login.