Moved getJsonBase code to waitForMsg function so that we can loop it
Move jQuery.ajax code to function waitForMsg and call it from getJsonBase.
Provide the parm repeat in getJsonBase so that calling functions can determine if they want the calls to be repeated periodically and the func they provide to handle the call output repeatedly.  We need this to make group.html.erb work, i.e., update the count periodically, without user clicking refresh.