Moved changes made to make test/integration/click_contact_to_send_contract_emails_test.rb work on new mac and corrected error with payments_controller.rb not passing payment_status to pay.html.erb when invoice_no is illegal.
email_elem.body is Mail::Body class in new setup to need to append .inspect.to_s in order to access it as a string.
Also in the original (old mac) payments_controller.rb, when invoice_no is illegal we forgot to update render "pay" to pass local constants such as payment_status.